feat: integration crud movies, sign-in, sign-up, activation email, private/public/admin route & state using redux

+++ src/helpers/axios.js
@@ -34,8 +34,20 @@ instances.interceptors.response.use(
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
     if (error.response.status === 403) {
-      localStorage.clear();
-      window.location.href = "/sign-in";
+      if (error.response.data.message === "jwt expired") {
+        const refreshToken = localStorage.getItem("refreshToken");
+        instances
+          .post("/auth/refresh-token", { refreshToken })
+          .then((res) => {
+            alert("Refresh token is success");
+            localStorage.setItem("token", res.value.data.data.token);
+            window.location.reload();
+          })
+          .catch((err) => {
+            localStorage.clear();
+            window.location.href = "/sign-in";
+          });
+      }
     }
     return Promise.reject(error);
   }
